+++ spec/avromatic/io/datum_writer_spec.rb
@@ -6,6 +6,7 @@ describe Avromatic::IO::DatumWriter do
   end
   let(:values) do
     {
+      header: 'foo',
       message: { bar_message: 'bar' }
     }
   end